+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Update MIBI directory schema
 - Update Visium no probes directory schema
 - Add Cell DIVE to index
+- Update Segmentation Masks directory schema
 
 ## v0.0.23
 - Add token to validation_utils.get_assaytype_data, replace URL string concatenation with urllib